+++ include/PowerTable_Helpers.h
@@ -89,6 +89,7 @@ class PTHelpers {
   ptIndex calculateIndex(int watts, int cad);
   void enterData(PTData& ptData,ptIndex index, int pos);
   void clean(PTData& ptData);
+  void fillGaps(PTData& ptData);
   bool fillAllWattColumns(PTData& ptData);
   bool fillAllCadenceLines(PTData& ptData);
   std::pair<std::vector<float>, std::vector<float>> getRow(int row, PTData& ptData);