+++ test/test.h
@@ -52,6 +52,13 @@ class TestTableFill {
     static void test_fill_incomplete_table(void);
 };
 
+class TestPAVAConvergence {
+  public:
+    static void test_pava_convergence_nebula3(void);
+    static void test_pava_convergence_nebula(void);
+    static void test_pava_convergence_better(void);
+};
+
 class TestAdevName2UniqueName {
 public:
     static void test_traditional_device_keeps_address_suffix(void);