+++ test/test_pt_lookup_resistance.cpp
@@ -24,7 +24,7 @@ void TestPTLookupResistance::test_pt_lookup_resistance(void) {
   PTData ptData;
 
   // Load the power table data from the .ptab file
-  const std::string filePath = "test/data/9077.ptab";
+  const std::string filePath = "test/data/Nebula3.ptab";
   loadCSVToPTData(filePath, ptData);
 
   // Create helpers object for lookup