+++ CHANGELOG.md
@@ -10,10 +10,27 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 ### Changed
+- Added feed forward, disabled PowerTable for ERG lookup. 
+- Added tests and removal of duplicates in pt column. 
+- Added removal of negative numbers in pt table. 
+
+### Hardware
+
+
+## [25.12.28]
+
+### Added
+
+### Changed
 - Removed >0 watts requirement to compute ERG.
 - Filter cadence for crazy values. Only >0 && <250 now accepted.
 - Filter watts for crazy values. Only >0 && <3000 now accepted. 
 - Fixed bug where scans may not happen even when configured devices aren't connected.  
+- Worked with Mark Roy to tune PID.
+- Added proper rounding from float to int for power and cadence.
+- More ERG tweaks for Marc Roy. 
+- If homed, we throw out negative PowerTable returns. 
+- After startup homing, set gear 8.
 
 ### Hardware